Move Dataset comparison tests to own file.
