+++ core/test/CMakeLists.txt
@@ -7,6 +7,7 @@ add_executable(${TARGET_NAME}
                data_array_test.cpp
                dataset_test.cpp
                dataset_test_common.cpp
+               dataset_comparison_test.cpp
                dataset_operations_test.cpp
                dimensions_test.cpp
                except_test.cpp
